Create README.md
Hello and welcome to Medivery! The fastest pharmacy in India! At medivery.in, we assist you in easily monitoring your own health as well as caring for loved ones wherever in India. With a few mouse clicks, you can buy and send drugs from anywhere in the country. What we do - We provide quick online access to medicines with simple home delivery at medivery.in. We make a comprehensive choice of prescription medicines and other health items easily accessible.